+++ test/e2e/app-dir/metadata-navigation/app/async/[slug]/page.tsx
