+++ test/e2e/app-dir/metadata-navigation/app/async/redirect/page.tsx
