+++ test/e2e/app-dir/metadata-thrown/app/layout.tsx
