+++ test/e2e/app-dir/metadata/app/dynamic/[slug]/page.tsx
@@ -1,0 +1,19 @@
+function format({ params, searchParams }) {
+  const { slug } = params
+  const { q } = searchParams
+  return `params - ${slug}${q ? ` query - ${q}` : ''}`
+}
+
+export default function page(props) {
+  return <p>{format(props)}</p>
+}
+
+export async function generateMetadata(props, parent) {
+  const parentMetadata = await parent
+  /* mutating */
+  return {
+    ...parentMetadata,
+    title: format(props),
+    keywords: parentMetadata.keywords.concat(['child']),
+  }
+}
