+++ test/e2e/app-dir/metadata/app/dynamic/layout.tsx
@@ -1,0 +1,11 @@
+export default function layout({ children }) {
+  return children
+}
+
+export async function generateMetadata() {
+  return {
+    keywords: 'parent',
+  }
+}
+
+export const revalidate = 0
