+++ test/lib/next-test-utils.ts
@@ -10,6 +10,7 @@ import { promisify } from 'util'
 import http from 'http'
 import path from 'path'
 
+import cheerio from 'cheerio'
 import spawn from 'cross-spawn'
 import { writeFile } from 'fs-extra'
 import getPort from 'get-port'
@@ -1232,3 +1233,150 @@ export const describeVariants = {
     }
   },
 }
+
+export const getTitle = (browser: BrowserInterface) =>
+  browser.elementByCss('title').text()
+
+async function checkMeta(
+  browser: BrowserInterface,
+  queryValue: string,
+  expected: RegExp | string | string[] | undefined | null,
+  queryKey: string = 'property',
+  tag: string = 'meta',
+  domAttributeField: string = 'content'
+) {
+  const values = await browser.eval(
+    `[...document.querySelectorAll('${tag}[${queryKey}="${queryValue}"]')].map((el) => el.getAttribute("${domAttributeField}"))`
+  )
+  if (expected instanceof RegExp) {
+    expect(values[0]).toMatch(expected)
+  } else {
+    if (Array.isArray(expected)) {
+      expect(values).toEqual(expected)
+    } else {
+      // If expected is undefined, then it should not exist.
+      // Otherwise, it should exist in the matched values.
+      if (expected === undefined) {
+        expect(values).not.toContain(undefined)
+      } else {
+        expect(values).toContain(expected)
+      }
+    }
+  }
+}
+
+export function createDomMatcher(browser: BrowserInterface) {
+  /**
+   * @param tag - tag name, e.g. 'meta'
+   * @param query - query string, e.g. 'name="description"'
+   * @param expectedObject - expected object, e.g. { content: 'my description' }
+   * @returns {Promise<void>} - promise that resolves when the check is done
+   *
+   * @example
+   * const matchDom = createDomMatcher(browser)
+   * await matchDom('meta', 'name="description"', { content: 'description' })
+   */
+  return async (
+    tag: string,
+    query: string,
+    expectedObject: Record<string, string | null | undefined>
+  ) => {
+    const props = await browser.eval(`
+      const el = document.querySelector('${tag}[${query}]');
+      const res = {}
+      const keys = ${JSON.stringify(Object.keys(expectedObject))}
+      for (const k of keys) {
+        res[k] = el?.getAttribute(k)
+      }
+      res
+    `)
+    expect(props).toEqual(expectedObject)
+  }
+}
+
+export function createMultiHtmlMatcher($: ReturnType<typeof cheerio.load>) {
+  /**
+   * @param tag - tag name, e.g. 'meta'
+   * @param queryKey - query key, e.g. 'property'
+   * @param domAttributeField - dom attribute field, e.g. 'content'
+   * @param expected - expected object, e.g. { description: 'my description' }
+   * @returns {void} - void when the check is done
+   *
+   * @example
+   *
+   * const $ = await next.render$('html')
+   * const matchHtml = createMultiHtmlMatcher($)
+   * matchHtml('meta', 'name', 'property', {
+   *   description: 'description',
+   *   og: 'og:description'
+   * })
+   *
+   */
+  return (
+    tag: string,
+    queryKey: string,
+    domAttributeField: string,
+    expected: Record<string, string | string[] | undefined>
+  ) => {
+    const res = {}
+    for (const key of Object.keys(expected)) {
+      const el = $(`${tag}[${queryKey}="${key}"]`)
+      if (el.length > 1) {
+        res[key] = el.toArray().map((el) => el.attribs[domAttributeField])
+      } else {
+        res[key] = el.attr(domAttributeField)
+      }
+    }
+    expect(res).toEqual(expected)
+  }
+}
+
+export function createMultiDomMatcher(browser: BrowserInterface) {
+  /**
+   * @param tag - tag name, e.g. 'meta'
+   * @param queryKey - query key, e.g. 'property'
+   * @param domAttributeField - dom attribute field, e.g. 'content'
+   * @param expected - expected object, e.g. { description: 'my description' }
+   * @returns {Promise<void>} - promise that resolves when the check is done
+   *
+   * @example
+   * const matchMultiDom = createMultiDomMatcher(browser)
+   * await matchMultiDom('meta', 'property', 'content', {
+   *   description: 'description',
+   *   'og:title': 'title',
+   *   'twitter:title': 'title'
+   * })
+   *
+   */
+  return async (
+    tag: string,
+    queryKey: string,
+    domAttributeField: string,
+    expected: Record<string, string | string[] | undefined | null>
+  ) => {
+    await Promise.all(
+      Object.keys(expected).map(async (key) => {
+        return checkMeta(
+          browser,
+          key,
+          expected[key],
+          queryKey,
+          tag,
+          domAttributeField
+        )
+      })
+    )
+  }
+}
+
+export const checkMetaNameContentPair = (
+  browser: BrowserInterface,
+  name: string,
+  content: string | string[]
+) => checkMeta(browser, name, content, 'name')
+
+export const checkLink = (
+  browser: BrowserInterface,
+  rel: string,
+  content: string | string[]
+) => checkMeta(browser, rel, content, 'rel', 'link', 'href')
